modify function distribution, and finish custom
