+++ lightx2v/models/networks/hunyuan/infer/feature_caching/transformer_infer_2.py
@@ -1,6 +1,5 @@
 from ..transformer_infer import BaseHunyuanTransformer
 import torch
-import numpy as np
 
 class HunyuanTransformerInferTeaCaching(BaseHunyuanTransformer):
     def __init__(self, config):
@@ -24,7 +23,7 @@ class HunyuanTransformerInferTeaCaching(BaseHunyuanTransformer):
         
         # 3. 更新数组
         if index <= self.scheduler.infer_steps - 2:
-            should_calc = self.calculate_should_calc(img, vec, weights)
+            should_calc = super().calculate_should_calc(img, vec, weights)
             self.scheduler.caching_records[index+1] = should_calc
 
         return img, vec
@@ -55,33 +54,3 @@ class HunyuanTransformerInferTeaCaching(BaseHunyuanTransformer):
     def infer_using_cache(self, weights, img, txt, vec, cu_seqlens_qkv, max_seqlen_qkv, freqs_cis, token_replace_vec=None, frist_frame_token_num=None):
         img += self.previous_residual
         return img, vec
-
-    # 1. only in tea-cache, judge next step
-    def calculate_should_calc(self, img, vec, weights):
-        # 1. 时间步嵌入调制
-        inp = img.clone()
-        vec_ = vec.clone()
-        img_mod1_shift, img_mod1_scale, _, _, _, _ = weights.double_blocks[0].img_mod.apply(vec_).chunk(6, dim=-1)
-        normed_inp = torch.nn.functional.layer_norm(inp, (inp.shape[1],), None, None, 1e-6)
-        modulated_inp = normed_inp * (1 + img_mod1_scale) + img_mod1_shift
-        del normed_inp, inp, vec_
-
-        # 2. L1距离计算
-        if self.scheduler.step_index == 0 or self.scheduler.step_index == self.scheduler.infer_steps - 1:
-            should_calc = True
-            self.accumulated_rel_l1_distance = 0
-        else:
-            rescale_func = np.poly1d(self.coefficients)
-            self.accumulated_rel_l1_distance += rescale_func(
-                ((modulated_inp -  self.previous_modulated_input).abs().mean() /  self.previous_modulated_input.abs().mean()).cpu().item()
-            )
-            if self.accumulated_rel_l1_distance < self.teacache_thresh:
-                should_calc = False
-            else:
-                should_calc = True
-                self.accumulated_rel_l1_distance = 0
-        self.previous_modulated_input = modulated_inp
-        del modulated_inp
-
-        # 3. 返回判断
-        return should_calc
