+++ lightx2v/models/networks/hunyuan/infer/feature_caching/transformer_infer_3.py
@@ -1,6 +1,5 @@
 from ..transformer_infer import BaseHunyuanTransformer
 import torch
-import numpy as np
 
 class HunyuanTransformerInferTaylorCaching(BaseHunyuanTransformer):
     def __init__(self, config):
@@ -24,17 +23,17 @@ class HunyuanTransformerInferTaylorCaching(BaseHunyuanTransformer):
         img_seq_len = img.shape[0]
         for i in range(self.double_blocks_num):
             img_out, txt_out, img_mod1_gate, img_mod2_shift, img_mod2_scale, img_mod2_gate,tr_img_mod1_gate, tr_img_mod2_shift, tr_img_mod2_scale, tr_img_mod2_gate, txt_mod1_gate, txt_mod2_shift, txt_mod2_scale, txt_mod2_gate = super().infer_double_block_1(weights.double_blocks[i], img, txt, vec, cu_seqlens_qkv, max_seqlen_qkv, freqs_cis, token_replace_vec, frist_frame_token_num)
-            self.derivative_approximation(self.double_blocks_cache[i], "img_attn", img_out)
-            self.derivative_approximation(self.double_blocks_cache[i], "txt_attn", txt_out)
+            super().derivative_approximation(self.double_blocks_cache[i], "img_attn", img_out)
+            super().derivative_approximation(self.double_blocks_cache[i], "txt_attn", txt_out)
             img, txt, img_out, txt_out, img_mod2_gate, txt_mod2_gate = super().infer_double_block_2(weights.double_blocks[i], img, txt, vec, cu_seqlens_qkv, max_seqlen_qkv, freqs_cis, token_replace_vec, frist_frame_token_num, img_out, txt_out, img_mod1_gate, img_mod2_shift, img_mod2_scale, img_mod2_gate,tr_img_mod1_gate, tr_img_mod2_shift, tr_img_mod2_scale, tr_img_mod2_gate, txt_mod1_gate, txt_mod2_shift, txt_mod2_scale, txt_mod2_gate)
-            self.derivative_approximation(self.double_blocks_cache[i],"img_mlp", img_out)
-            self.derivative_approximation(self.double_blocks_cache[i],"txt_mlp", txt_out)
+            super().derivative_approximation(self.double_blocks_cache[i],"img_mlp", img_out)
+            super().derivative_approximation(self.double_blocks_cache[i],"txt_mlp", txt_out)
             img, txt = super().infer_double_block_3(weights.double_blocks[i], img, txt, vec, cu_seqlens_qkv, max_seqlen_qkv, freqs_cis, token_replace_vec, frist_frame_token_num, img_out, txt_out, img_mod2_gate, txt_mod2_gate)
 
         x = torch.cat((img, txt), 0)
         for i in range(self.single_blocks_num):
             out, mod_gate, tr_mod_gate = super().infer_single_block_1(weights.single_blocks[i], x, vec, txt_seq_len, cu_seqlens_qkv, max_seqlen_qkv, freqs_cis, token_replace_vec, frist_frame_token_num)
-            self.derivative_approximation(self.single_blocks_cache[i], "total", out)
+            super().derivative_approximation(self.single_blocks_cache[i], "total", out)
             x = super().infer_single_block_2(weights.single_blocks[i], x, vec, txt_seq_len, cu_seqlens_qkv, max_seqlen_qkv, freqs_cis, out, mod_gate, tr_mod_gate, token_replace_vec, frist_frame_token_num)
         img = x[:img_seq_len, ...]
         return img, vec
@@ -52,18 +51,7 @@ class HunyuanTransformerInferTaylorCaching(BaseHunyuanTransformer):
         img = x[:img_seq_len, ...]
         return img, vec
     
-    # 1. when fully calcualted, stored in cache
-    def derivative_approximation(self, block_cache, module_name, out):
-        if module_name not in block_cache:
-            block_cache[module_name] = {0: out}
-        else:
-            step_diff = super().get_taylor_step_diff()
-
-            previous_out = block_cache[module_name][0]
-            block_cache[module_name][0] = out
-            block_cache[module_name][1] = (out - previous_out) / step_diff
-
-    # 2. taylor using caching
+    # 1. taylor using caching
     def infer_double_block(self, weights, img, txt, vec, cu_seqlens_qkv, max_seqlen_qkv, freqs_cis, i):
         vec_silu = torch.nn.functional.silu(vec)
         img_mod_out = weights.img_mod.apply(vec_silu)
