+++ lightx2v/models/networks/hunyuan/infer/transformer_infer.py
@@ -2,6 +2,7 @@ from lightx2v.models.networks.aaa_transformer.transformer_infer import BaseTrans
 import torch
 from einops import rearrange
 from .utils_bf16 import apply_rotary_emb
+import numpy as np
 
 class BaseHunyuanTransformer(BaseTransformer):
     # 1. 初始化
@@ -215,3 +216,44 @@ class BaseHunyuanTransformer(BaseTransformer):
             out = out * mod_gate
         x = x + out
         return x
+
+    # 1. only in tea-cache, judge next step
+    def calculate_should_calc(self, img, vec, weights):
+        # 1. 时间步嵌入调制
+        inp = img.clone()
+        vec_ = vec.clone()
+        img_mod1_shift, img_mod1_scale, _, _, _, _ = weights.double_blocks[0].img_mod.apply(vec_).chunk(6, dim=-1)
+        normed_inp = torch.nn.functional.layer_norm(inp, (inp.shape[1],), None, None, 1e-6)
+        modulated_inp = normed_inp * (1 + img_mod1_scale) + img_mod1_shift
+        del normed_inp, inp, vec_
+
+        # 2. L1距离计算
+        if self.scheduler.step_index == 0 or self.scheduler.step_index == self.scheduler.infer_steps - 1:
+            should_calc = True
+            self.accumulated_rel_l1_distance = 0
+        else:
+            rescale_func = np.poly1d(self.coefficients)
+            self.accumulated_rel_l1_distance += rescale_func(
+                ((modulated_inp -  self.previous_modulated_input).abs().mean() /  self.previous_modulated_input.abs().mean()).cpu().item()
+            )
+            if self.accumulated_rel_l1_distance < self.teacache_thresh:
+                should_calc = False
+            else:
+                should_calc = True
+                self.accumulated_rel_l1_distance = 0
+        self.previous_modulated_input = modulated_inp
+        del modulated_inp
+
+        # 3. 返回判断
+        return should_calc
+    
+    # 1. when fully calcualted, stored in cache
+    def derivative_approximation(self, block_cache, module_name, out):
+        if module_name not in block_cache:
+            block_cache[module_name] = {0: out}
+        else:
+            step_diff = super().get_taylor_step_diff()
+
+            previous_out = block_cache[module_name][0]
+            block_cache[module_name][0] = out
+            block_cache[module_name][1] = (out - previous_out) / step_diff
