+++ lightx2v/models/schedulers/hunyuan/feature_caching/scheduler.py
@@ -36,3 +36,11 @@ class HunyuanSchedulerAdaCaching(HunyuanScheduler):
     def clear(self):
         # TODO: Transformer实例的缓存清理
         pass
+
+class HunyuanSchedulerCustomCaching(HunyuanScheduler):
+    def __init__(self, config):
+        super().__init__(config)
+
+    def clear(self):
+        # TODO: Transformer实例的缓存清理
+        pass
